requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
asttokens=2.2.1=pypi_0
backcall=0.2.0=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.01.10=h06a4308_0
cachetools=5.3.1=pypi_0
certifi=2023.5.7=pypi_0
charset-normalizer=3.1.0=pypi_0
cmake=3.26.3=pypi_0
contourpy=1.0.7=pypi_0
cycler=0.11.0=pypi_0
cython=3.0.0=pypi_0
decorator=5.1.1=pypi_0
executing=1.2.0=pypi_0
filelock=3.12.0=pypi_0
fonttools=4.39.4=pypi_0
google-auth=2.19.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.54.2=pypi_0
idna=3.4=pypi_0
ipython=8.13.2=pypi_0
jedi=0.18.2=pypi_0
jinja2=3.1.2=pypi_0
kiwisolver=1.4.4=pypi_0
lapx=0.5.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
lit=16.0.5=pypi_0
markdown=3.4.3=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.7.1=pypi_0
matplotlib-inline=0.1.6=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.1=pypi_0
numpy=1.24.3=pypi_0
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-cupti-cu11=11.7.101=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.2.10.91=pypi_0
nvidia-cusolver-cu11=11.4.0.1=pypi_0
nvidia-cusparse-cu11=11.7.4.91=pypi_0
nvidia-nccl-cu11=2.14.3=pypi_0
nvidia-nvtx-cu11=11.7.91=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.7.0.72=pypi_0
openssl=1.1.1t=h7f8727e_0
packaging=23.1=pypi_0
pandas=2.0.2=pypi_0
parso=0.8.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.5.0=pypi_0
pip=23.0.1=pypi_0
prompt-toolkit=3.0.38=pypi_0
protobuf=4.21.2=pypi_0
psutil=5.9.5=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pygments=2.15.1=pypi_0
pyparsing=3.0.9=pypi_0
python=3.10.11=h7a1cb2a_2
python-dateutil=2.8.2=pypi_0
pytz=2023.3=pypi_0
pyyaml=6.0=pypi_0
readline=8.2=h5eee18b_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.12.2=pypi_0
setuptools=67.8.0=pypi_0
six=1.16.0=pypi_0
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.2=pypi_0
sympy=1.12=pypi_0
tensorboard=2.13.0=pypi_0
tensorboard-data-server=0.7.0=pypi_0
thop=0.1.1-2209072238=pypi_0
tk=8.6.12=h1ccaba5_0
torch=2.0.1=pypi_0
torchvision=0.15.2=pypi_0
tqdm=4.65.0=pypi_0
traitlets=5.9.0=pypi_0
triton=2.0.0=pypi_0
typing-extensions=4.6.2=pypi_0
tzdata=2023.3=pypi_0
ultralytics=8.0.143=pypi_0
urllib3=1.26.16=pypi_0
wcwidth=0.2.6=pypi_0
werkzeug=2.3.4=pypi_0
wheel=0.38.4=pypi_0
xz=5.4.2=h5eee18b_0
zlib=1.2.13=h5eee18b_0
